Renewed the supporting_functions.py and clarified how to set grid resolutions
I changed the `supporting_functions.py` at line 403 to:

```
if grid_res == 0:
        avg_grav_spacing, _ = calculate_average_gravity_spacing(grav_data_loc, method='2d', use_z=False)
        grid_res = 50*np.round(avg_grav_spacing/100)
    elif grid_res < 0:
        raise ValueError('Positive grid resolutions only!')
```

To give an error message when a negative value is passed to `grid_res`. The default value of `grid_res` is given when `grid_res=0` and is `50*np.round(average_grav_spacing/100)`. Otherwise, the function will take the `grid_res` value as it is.

In the `Equivalent Surfaces.ipynb` and the rendered HTML, I clarified how to set the grid resolution using `grid_res` for gridding the gravity data at a planar surface above the topography.